Created a series of functions in which the stan_jm arguments are checked
The functions were called to check that the user had appropriately
entered each of the arguments to stan_jm. They aren't ideal though, so I
don't think these changes will be pulled into the master stan_jm.
Rather, it is probably better to retain error checks as part of the main
code for the stan_jm modelling function.