Added survival predictions and plots of survival function
posterior_survfit now calculates subject-specific and marginalised
survival predictions based on draws from the posterior. This function
returns an object of class "survfit.stanjm" for which I have added a
plot method. By default the function extrapolates beyond the last
observation (measurement, event or censoring) time for an individual --
the distance extrapolated is an additional 50% of the observed follow up
time for that individual.