Fixed two bugs, detailed below
Made sure that model frame time variable is numeric before data.table
merge. It was converting the quadrature times to integers if the time
variable in the original model frame was integer and not numeric.

Also fixed the Zqmerge and Zqmerge_eps design matrices in stan_jm so
that they are subjected to the pad_reTrms function, just like the Z
design matrix is. This ensures that the matrices are of the correct size
before passing them to Stan. They were the wrong size before this bug
fix, and therefore the multivariate joint model was returning incorrect
estimates for the second marker.