Significant changes to posterior_predict
Removed posterior_traj. Consolidated everything into posterior_predict.
Simplified the number of arguments by including a "control" list. Also,
allowed the user to specify whether a matrix or data frame should be
returned. Now uses a

Also updated the documentation for posterior_predict, so that it should
now be complete, with examples.