Added options to interact association structures with data/covariates
At the moment the data must be contained in the dataframe used for the
covariates in the longitudinal submodel. However I have also added an
argument to stan_jm that allows the user to specify "interaction_data",
which could be a data frame that contains the covariates but using a
different measurement time schedule to the data frame passed to
dataLong. This would provide greater flexibility. The "interaction_data"
argument isn't properly implemented yet though.